Move TDP components to `web/shared` (#53501)
* Move `teleport/ironrdp` to `shared/libs/ironrdp`

* Do not import `teleport/lib/AuthenticatedWebSocket` in `TdpClient`

* Do not import other `teleport` stuff in `TdpClient`

* Move `teleport/lib/tdp` to `shared/libs/tdp`

* Move `teleport/components/TdpClientCanvas` to `shared/components/CanvasRenderer`

* Move `teleport/DesktopSession` to `shared/components/DesktopSession`

* Bring back `teleport/DesktopSession`

* Improve old artifacts cleaning

* Fix path in README.md

(cherry picked from commit d272c038202fe663ffd5b7326cde7edcf324693c)
