+++ web/packages/shared/README.md
@@ -11,3 +11,11 @@ Add `@gravitational/shared` to your package.json file.
     "@gravitational/shared": "^1.0.0",
   },
 ```
+
+### WASM
+
+This package includes a WASM module built from a Rust codebase located in `packages/shared/libs/ironrdp`.
+It is built with the help of [wasm-pack](https://github.com/rustwasm/wasm-pack).
+
+Running `pnpm build-wasm` builds the WASM binary as well as the appropriate Javascript/Typescript
+bindings and types in `web/packages/shared/libs/ironrdp/pkg`.
