+++ web/packages/shared/components/DesktopSession/TopBar.tsx
@@ -37,8 +37,8 @@ export default function TopBar(props: Props) {
     isSharingDirectory,
     onShareDirectory,
     onCtrlAltDel,
-    warnings,
-    onRemoveWarning,
+    alerts,
+    onRemoveAlert,
   } = props;
   const theme = useTheme();
 
@@ -79,10 +79,7 @@ export default function TopBar(props: Props) {
           >
             <Clipboard style={primaryOnTrue(isSharingClipboard)} pr={3} />
           </HoverTooltip>
-          <WarningDropdown
-            warnings={warnings}
-            onRemoveWarning={onRemoveWarning}
-          />
+          <WarningDropdown warnings={alerts} onRemoveWarning={onRemoveAlert} />
         </Flex>
         <ActionMenu
           onDisconnect={onDisconnect}
@@ -117,6 +114,6 @@ type Props = {
   onDisconnect: VoidFunction;
   onShareDirectory: VoidFunction;
   onCtrlAltDel: VoidFunction;
-  warnings: NotificationItem[];
-  onRemoveWarning(id: string): void;
+  alerts: NotificationItem[];
+  onRemoveAlert(id: string): void;
 };
