+++ web/packages/shared/components/DesktopSession/useDesktopSession.tsx
@@ -27,10 +27,7 @@ import {
 
 import type { NotificationItem } from 'shared/components/Notification';
 import { Attempt } from 'shared/hooks/useAsync';
-
-import { TdpClient } from 'teleport/lib/tdp';
-import { ClipboardData } from 'teleport/lib/tdp/codec';
-import { Sha256Digest } from 'teleport/lib/util';
+import { ClipboardData, TdpClient } from 'shared/libs/tdp';
 
 declare global {
   interface Window {
@@ -102,7 +99,7 @@ export default function useDesktopSession(
       return;
     }
     const text = await navigator.clipboard.readText();
-    const digest = await Sha256Digest(text, encoder.current);
+    const digest = await sha256Digest(text, encoder.current);
     if (text && digest !== latestClipboardDigest.current) {
       tdpClient.sendClipboardData({
         data: text,
@@ -117,7 +114,7 @@ export default function useDesktopSession(
       (await sysClipboardGuard(clipboardSharingState, 'write'))
     ) {
       await navigator.clipboard.writeText(clipboardData.data);
-      latestClipboardDigest.current = await Sha256Digest(
+      latestClipboardDigest.current = await sha256Digest(
         clipboardData.data,
         encoder.current
       );
@@ -383,4 +380,16 @@ async function sysClipboardGuard(
   // but the other is denied, and we want to try the operation. The feature is
   // either fully enabled or fully disabled.
   return isSharingClipboard(clipboardSharingState) && document.hasFocus();
+}
+
+// Adapted from https://developer.mozilla.org/en-US/docs/Web/API/SubtleCrypto/digest#converting_a_digest_to_a_hex_string
+async function sha256Digest(
+  message: string,
+  encoder: TextEncoder = new TextEncoder()
+): Promise<string> {
+  const msgUint8 = encoder.encode(message); // encode as (utf-8) Uint8Array
+  const hashBuffer = await crypto.subtle.digest('SHA-256', msgUint8); // hash the message
+  const hashArray = Array.from(new Uint8Array(hashBuffer)); // convert buffer to byte array
+  const hashHex = hashArray.map(b => b.toString(16).padStart(2, '0')).join(''); // convert bytes to hex string
+  return hashHex;
 }
