+++ web/packages/shared/libs/tdp/index.ts
@@ -16,4 +16,10 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export { DesktopSession as DesktopSessionContainer } from './DesktopSession';
+export {
+  TdpClient,
+  TdpClientEvent,
+  useListener,
+  type BitmapFrame,
+} from './client';
+export * from './codec';
