+++ web/packages/teleport/src/Player/DesktopPlayer.tsx
@@ -26,13 +26,15 @@ import React, {
 import styled from 'styled-components';
 
 import { Alert, Box, Flex, Indicator } from 'design';
+import {
+  CanvasRenderer,
+  CanvasRendererRef,
+} from 'shared/components/CanvasRenderer';
+import { useListener } from 'shared/libs/tdp';
 
-import TdpClientCanvas from 'teleport/components/TdpClientCanvas';
-import { TdpClientCanvasRef } from 'teleport/components/TdpClientCanvas/TdpClientCanvas';
 import cfg from 'teleport/config';
 import { formatDisplayTime, StatusEnum } from 'teleport/lib/player';
 import { PlayerClient } from 'teleport/lib/tdp';
-import { useListener } from 'teleport/lib/tdp/client';
 import { getHostName } from 'teleport/services/api';
 
 import ProgressBar from './ProgressBar';
@@ -61,7 +63,7 @@ export const DesktopPlayer = ({
     sid,
     clusterId,
   });
-  const tdpClientCanvasRef = useRef<TdpClientCanvasRef>(null);
+  const canvasRendererRef = useRef<CanvasRendererRef>(null);
 
   useListener(playerClient?.onError, clientOnTdpError);
   useListener(playerClient?.onClientError, clientOnTdpError);
@@ -70,15 +72,15 @@ export const DesktopPlayer = ({
   useListener(playerClient?.onWsClose, clientOnWsClose);
   useListener(
     playerClient?.onPngFrame,
-    tdpClientCanvasRef.current?.renderPngFrame
+    canvasRendererRef.current?.renderPngFrame
   );
   useListener(
     playerClient?.onBmpFrame,
-    tdpClientCanvasRef.current?.renderBitmapFrame
+    canvasRendererRef.current?.renderBitmapFrame
   );
   useListener(
     playerClient?.onScreenSpec,
-    tdpClientCanvasRef.current?.setResolution
+    canvasRendererRef.current?.setResolution
   );
 
   const isError = playerStatus === StatusEnum.ERROR || statusText !== '';
@@ -100,7 +102,7 @@ export const DesktopPlayer = ({
       )}
 
       <StyledContainer>
-        <TdpClientCanvas ref={tdpClientCanvasRef} />
+        <CanvasRenderer ref={canvasRendererRef} />
 
         <ProgressBar
           min={0}
